requirements.txt
python-telegram-bot>=20.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytz>=2024.1
